react/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
snowflake-connector-python>=3.0.0
pydantic>=2.0.0
httpx>=0.25.0
